refactor tests and deleted unnecessary tests
